requirements.txt
# Core LangChain
langchain>=0.3.25
langchain-core>=0.3.63
langchain-community>=0.3.24
langchain-text-splitters>=0.3.8

# Integraciones específicas
langchain-ollama==0.2.0
langchain-chroma==0.1.4

# PDF → Markdown (herramienta auxiliar, no requerido para ingesta principal)
PyMuPDF>=1.24.0

# Utilidades
python-dotenv==1.0.1
reportlab>=4.0.0

# UI
streamlit==1.38.0

# Tests
pytest>=7.4.0

# Evaluación RAG (barras de progreso)
tqdm>=4.66.0

# Orquestación (grafo determinista F4)
langgraph>=0.2.14,<0.3
